+++ src/Admin/BookmarkAdmin.php
@@ -1,0 +1,82 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Admin;
+
+use Doctrine\ORM\QueryBuilder;
+use Sonata\AdminBundle\Admin\AbstractAdmin;
+use Sonata\AdminBundle\Datagrid\DatagridMapper;
+use Sonata\AdminBundle\Datagrid\ListMapper;
+use Sonata\AdminBundle\Form\FormMapper;
+use Sonata\AdminBundle\Show\ShowMapper;
+
+final class BookmarkAdmin extends AbstractAdmin
+{
+    protected $datagridValues = [
+        '_page' => 1,
+        '_sort_order' => 'DESC',
+        '_sort_by' => 'created',
+    ];
+
+    protected function configureDatagridFilters(DatagridMapper $datagridMapper): void
+    {
+        $datagridMapper
+            ->add('id')
+            ->add('title')
+            ->add('timeInSeconds')
+            ->add('lesson')
+            ->add('created')
+            ->add('updated')
+            ;
+    }
+
+    protected function configureListFields(ListMapper $listMapper): void
+    {
+        $listMapper
+            ->add('id')
+            ->add('title')
+            ->add('timeInSeconds')
+            ->add('lesson')
+            ->add('created')
+            ->add('updated')
+            ->add('_action', null, [
+                'actions' => [
+                    'show' => [],
+                    'edit' => [],
+                    'delete' => [],
+                ],
+            ]);
+    }
+
+    protected function configureFormFields(FormMapper $formMapper): void
+    {
+        $formMapper
+            ->add('title')
+            ->add('timeInSeconds')
+            ->add('lesson')
+            ;
+    }
+
+    protected function configureShowFields(ShowMapper $showMapper): void
+    {
+        $showMapper
+            ->add('id')
+            ->add('title')
+            ->add('timeInSeconds')
+            ->add('created')
+            ->add('updated')
+            ;
+    }
+
+    public function createQuery($context = 'list')
+    {
+        /** @var QueryBuilder $query */
+        $query = parent::createQuery($context);
+        $query->andWhere(
+            $query->expr()->isnull($query->getRootAliases()[0].'.user')
+        );
+
+        return  $query;
+    }
+}