+++ src/Form/BookmarkType.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace App\Form;
+
+use App\Entity\Bookmark;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\NumberType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class BookmarkType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('title', TextType::class, [
+                'required' => true,
+            ])
+            ->add('timeInSeconds', NumberType::class, [
+                'required' => true,
+            ])
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => Bookmark::class,
+            'csrf_protection' => false,
+        ]);
+    }
+}