+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - [Unreleased]
 
 ### Added
+- Lesson bookmarks feature. Users cane see admin defined bookmarks and create/delete their own.
 - Embed type field. Can be used as an extra information about embed handling by students app. 
 - New Author property - gender (it will allow fix prononciation in author box in students app)
 - API for listing all course attachments
 - User password reset request and reset actions
 
-[Unreleased]: https://github.com/owncourses/courses-server/compare/cbe9d0b142e7c6ace6b392412941075990494a47...41960407e7a68ba7d867509108d51de4d78618a2
+[Unreleased]: https://github.com/owncourses/courses-server/compare/41960407e7a68ba7d867509108d51de4d78618a2...HEAD
+
+[Proof Of Concept work]: https://github.com/owncourses/courses-server/compare/cbe9d0b142e7c6ace6b392412941075990494a47...41960407e7a68ba7d867509108d51de4d78618a2